requirements.txt
customtkinter>=5.2.2
pyserial>=3.5
requests>=2.32
